+++ src/games/cclcc/tipssystem.cpp
@@ -60,25 +60,25 @@ void TipsSystem::UpdateTipRecords() {
   }
 }
 
-void TipsSystem::SetTipLockedState(int id, bool state) {
+void TipsSystem::SetTipLockedState(size_t id, bool state) {
   Records[id].IsLocked = state;
   SaveSystem::SetTipStatus(id, Records[id].IsLocked, Records[id].IsUnread,
                            Records[id].IsNew);
 }
 
-void TipsSystem::SetTipUnreadState(int id, bool state) {
+void TipsSystem::SetTipUnreadState(size_t id, bool state) {
   Records[id].IsUnread = state;
   SaveSystem::SetTipStatus(id, Records[id].IsLocked, Records[id].IsUnread,
                            Records[id].IsNew);
 }
 
-void TipsSystem::SetTipNewState(int id, bool state) {
+void TipsSystem::SetTipNewState(size_t id, bool state) {
   Records[id].IsNew = state;
   SaveSystem::SetTipStatus(id, Records[id].IsLocked, Records[id].IsUnread,
                            Records[id].IsNew);
 }
 
-bool TipsSystem::GetTipLockedState(int id) { return Records[id].IsLocked; }
+bool TipsSystem::GetTipLockedState(size_t id) { return Records[id].IsLocked; }
 
 }  // namespace CCLCC
 }  // namespace Impacto
